+++ lib/arcanus.rb
@@ -37,13 +37,22 @@ module Arcanus
             'Run `arcanus setup`'
     end
 
-    unless File.exist?(@repo.unlocked_key_path)
+    if File.exist?(@repo.unlocked_key_path)
+      key = Arcanus::Key.from_file(@repo.locked_key_path)
+    elsif ENV['ARCANUS_PASSWORD']
+      key = Arcanus::Key.from_protected_file(@repo.locked_key_path, ENV['ARCANUS_PASSWORD'])
+      ENV.delete('ARCANUS_PASSWORD') # Scrub so child processes don't inherit
+    else
       raise Errors::UsageError,
             'Arcanus key has not been unlocked. ' \
-            'Run `arcanus unlock`'
+            'Run `arcanus unlock` or specify password via ARCANUS_PASSWORD environment variable'
     end
 
-    @chest = Chest.new(key_file_path: @repo.unlocked_key_path,
-                       chest_file_path: @repo.chest_file_path)
+    @chest = Chest.new(key: key, chest_file_path: @repo.chest_file_path)
+  end
+
+  def unlock(password)
+    key = Arcanus::Key.from_protected_file(@repo.locked_key_path, password)
+    key.save(key_file_path: @repo.unlocked_key_path)
   end
 end